@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.1 <3.0.0"
+  sdk: ">=2.19.2 <3.0.0"
 
 dependencies:
   flutter:
@@ -28,34 +28,38 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
-  google_sign_in: ^5.0.4
+  google_sign_in: ^6.1.5
   percent_indicator: ^4.2.2
-  cached_network_image: ^3.0.0
   flutter_spinkit: ^5.0.0
-  cloud_firestore: ^3.4.3
+  cloud_firestore: ^4.9.1
   fluttertoast: ^8.0.7
   animations: ^2.0.0
-  image: ^3.0.2
-  uuid: ^3.0.4
-  firebase_auth: ^3.6.2
-  firebase_core: ^1.3.0
-  image_picker: ^0.8.1+1
-  firebase_storage: ^10.3.4
+  image: ^4.0.17
+  uuid: ^4.0.0
+  firebase_auth: ^4.9.0
+  firebase_core: ^2.15.1
+  image_picker: ^1.0.4
+  firebase_storage: ^11.2.6
   path_provider: ^2.0.2
-  geolocator: ^9.0.1
-  geocoding: ^2.0.0
+  geolocator: ^10.0.1
   timeago: ^3.1.0
-  google_fonts: ^3.0.1
+  google_fonts: ^6.0.0
   provider: ^6.0.3
-  image_cropper: ^3.0.0
+  image_cropper: ^5.0.0
   shared_preferences: ^2.0.6
   ionicons: ^0.2.1
   loading_overlay: ^0.3.0
   like_button: ^2.0.4
   flutter_chat_bubble: ^2.0.0
-  permission_handler: ^10.1.0
-  story: ^0.8.0
-  story_view: ^0.14.0
+  permission_handler: ^11.0.0
+  story: ^1.1.0
+  story_view: ^0.15.0
+  iconsax: ^0.0.8
+  geocoding: ^2.0.0
+  just_audio: ^0.9.35
+  shimmer: ^3.0.0
+  video_player: ^2.7.2
+  cached_network_image: ^3.3.0
 
 dev_dependencies:
   flutter_test:
