@@ -1,7 +1,10 @@
 import 'package:animations/animations.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:iconsax/iconsax.dart';
+import 'package:ionicons/ionicons.dart';
 import 'package:provider/provider.dart';
+import 'package:social_media_app/posts/create_reel.dart';
 import 'package:social_media_app/posts/story/confrim_status.dart';
 import 'package:social_media_app/view_models/status/status_view_model.dart';
 import '../posts/create_post.dart';
@@ -52,7 +55,7 @@ class FabContainer extends StatelessWidget {
       ),
       builder: (BuildContext context) {
         return FractionallySizedBox(
-          heightFactor: .4,
+          heightFactor: .55,
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
@@ -73,7 +76,7 @@ class FabContainer extends StatelessWidget {
               Divider(),
               ListTile(
                 leading: Icon(
-                  CupertinoIcons.camera_on_rectangle,
+                  CupertinoIcons.photo_fill_on_rectangle_fill,
                   size: 25.0,
                 ),
                 title: Text('Make a post'),
@@ -88,14 +91,28 @@ class FabContainer extends StatelessWidget {
               ),
               ListTile(
                 leading: Icon(
-                  CupertinoIcons.camera_on_rectangle,
+                  CupertinoIcons.camera_circle,
                   size: 25.0,
                 ),
                 title: Text('Add to story'),
                 onTap: () async {
                   // Navigator.pop(context);
                   await viewModel.pickImage(context: context);
-
+                },
+              ),
+              ListTile(
+                leading: Icon(
+                  Iconsax.video_square,
+                  size: 25.0,
+                ),
+                title: Text('Post a reel'),
+                onTap: () async {
+                  Navigator.pop(context);
+                  Navigator.of(context).push(
+                    CupertinoPageRoute(
+                      builder: (_) => CreateReel(),
+                    ),
+                  );
                 },
               ),
             ],
