@@ -1,11 +1,17 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/scheduler.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:iconsax/iconsax.dart';
 import 'package:ionicons/ionicons.dart';
+import 'package:provider/provider.dart';
 import 'package:social_media_app/chats/recent_chats.dart';
 import 'package:social_media_app/models/post.dart';
+import 'package:social_media_app/screens/notification.dart';
 import 'package:social_media_app/utils/constants.dart';
 import 'package:social_media_app/utils/firebase.dart';
+import 'package:social_media_app/view_models/status/home/home_view_model.dart';
 import 'package:social_media_app/widgets/indicators.dart';
 import 'package:social_media_app/widgets/story_widget.dart';
 import 'package:social_media_app/widgets/userpost.dart';
@@ -15,30 +21,22 @@ class Feeds extends StatefulWidget {
   _FeedsState createState() => _FeedsState();
 }
 
-class _FeedsState extends State<Feeds> with AutomaticKeepAliveClientMixin{
+class _FeedsState extends State<Feeds> with AutomaticKeepAliveClientMixin {
   final GlobalKey<ScaffoldState> scaffoldKey = GlobalKey<ScaffoldState>();
-
-  int page = 5;
-  bool loadingMore = false;
-  ScrollController scrollController = ScrollController();
 
   @override
   void initState() {
-    scrollController.addListener(() async {
-      if (scrollController.position.pixels ==
-          scrollController.position.maxScrollExtent) {
-        setState(() {
-          page = page + 5;
-          loadingMore = true;
-        });
-      }
-    });
+    SchedulerBinding.instance.addPostFrameCallback(
+      (_) {
+        Provider.of<HomeViewModel>(context, listen: false).fetchFeeds();
+      },
+    );
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
-    print('>>>');
+    HomeViewModel viewModel = Provider.of<HomeViewModel>(context);
     return Scaffold(
       key: scaffoldKey,
       appBar: AppBar(
@@ -46,6 +44,7 @@ class _FeedsState extends State<Feeds> with AutomaticKeepAliveClientMixin{
         title: Text(
           Constants.appName,
           style: TextStyle(
+            fontSize: 25.0,
             fontWeight: FontWeight.w900,
           ),
         ),
@@ -53,8 +52,22 @@ class _FeedsState extends State<Feeds> with AutomaticKeepAliveClientMixin{
         actions: [
           IconButton(
             icon: Icon(
-              Ionicons.chatbubble_ellipses,
-              size: 30.0,
+              Iconsax.notification,
+              size: 25.0,
+            ),
+            onPressed: () {
+              Navigator.push(
+                context,
+                CupertinoPageRoute(
+                  builder: (_) => Activities(),
+                ),
+              );
+            },
+          ),
+          IconButton(
+            icon: Icon(
+              Iconsax.message,
+              size: 25.0,
             ),
             onPressed: () {
               Navigator.push(
@@ -65,61 +78,40 @@ class _FeedsState extends State<Feeds> with AutomaticKeepAliveClientMixin{
               );
             },
           ),
-          SizedBox(width: 20.0),
+          SizedBox(width: 5.0),
         ],
       ),
       body: RefreshIndicator(
         color: Theme.of(context).colorScheme.secondary,
-        onRefresh: () =>
-            postRef.orderBy('timestamp', descending: true).limit(page).get(),
-        child: SingleChildScrollView(
-          // controller: scrollController,
-          physics: NeverScrollableScrollPhysics(),
+        onRefresh: () => viewModel.fetchFeeds(),
+        child: Container(
+          height: MediaQuery.of(context).size.height,
+          width: MediaQuery.of(context).size.width,
           child: Column(
             mainAxisAlignment: MainAxisAlignment.start,
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
               StoryWidget(),
-              Container(
-                height: MediaQuery.of(context).size.height,
-                child: FutureBuilder(
-                  future: postRef
-                      .orderBy('timestamp', descending: true)
-                      .limit(page)
-                      .get(),
-                  builder: (context, AsyncSnapshot<QuerySnapshot> snapshot) {
-                    if (snapshot.hasData) {
-                      var snap = snapshot.data;
-                      List docs = snap!.docs;
-                      return ListView.builder(
-                        controller: scrollController,
-                        itemCount: docs.length,
-                        shrinkWrap: true,
-                        itemBuilder: (context, index) {
-                          PostModel posts =
-                              PostModel.fromJson(docs[index].data());
-                          return Padding(
-                            padding: const EdgeInsets.all(10.0),
-                            child: UserPost(post: posts),
-                          );
-                        },
-                      );
-                    } else if (snapshot.connectionState ==
-                        ConnectionState.waiting) {
-                      return circularProgress(context);
-                    } else
-                      return Center(
-                        child: Text(
-                          'No Feeds',
-                          style: TextStyle(
-                            fontSize: 26.0,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                      );
+              Flexible(
+                child: ListView.builder(
+                  controller: viewModel.scrollController,
+                  itemCount: viewModel.data.length,
+                  shrinkWrap: true,
+                  itemBuilder: (context, index) {
+                    PostModel posts = PostModel.fromJson(
+                      viewModel.data[index].data() as Map<String, dynamic>,
+                    );
+                    return Padding(
+                      padding: const EdgeInsets.all(10.0),
+                      child: UserPost(post: posts),
+                    );
                   },
                 ),
               ),
+              if (viewModel.loadingMore)
+                Center(
+                  child: CupertinoActivityIndicator(),
+                )
             ],
           ),
         ),
