@@ -1,11 +1,14 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
+import 'package:iconsax/iconsax.dart';
 import 'package:ionicons/ionicons.dart';
 import 'package:loading_overlay/loading_overlay.dart';
 import 'package:provider/provider.dart';
 import 'package:social_media_app/components/custom_image.dart';
 import 'package:social_media_app/models/user.dart';
+import 'package:social_media_app/utils/constants.dart';
 import 'package:social_media_app/utils/firebase.dart';
 import 'package:social_media_app/view_models/auth/posts_view_model.dart';
 import 'package:social_media_app/widgets/indicators.dart';
@@ -41,21 +44,27 @@ class _CreatePostState extends State<CreatePost> {
                 Navigator.pop(context);
               },
             ),
-            title: Text('WOOBLE'.toUpperCase()),
+            title: Text(
+              Constants.appName,
+              style: TextStyle(
+                fontSize: 25.0,
+                fontWeight: FontWeight.w900,
+              ),
+            ),
             centerTitle: true,
             actions: [
               GestureDetector(
                 onTap: () async {
                   await viewModel.uploadPosts(context);
-                  Navigator.pop(context);
-                  viewModel.resetPost();
+                  // Navigator.pop(context);
+                  // viewModel.resetPost();
                 },
                 child: Padding(
                   padding: const EdgeInsets.all(20.0),
                   child: Text(
                     'Post'.toUpperCase(),
                     style: TextStyle(
-                      fontWeight: FontWeight.bold,
+                      fontWeight: FontWeight.w900,
                       fontSize: 15.0,
                       color: Theme.of(context).colorScheme.secondary,
                     ),
@@ -77,7 +86,7 @@ class _CreatePostState extends State<CreatePost> {
                     );
                     return ListTile(
                       leading: CircleAvatar(
-                        radius: 25.0,
+                        radius: 30.0,
                         backgroundImage: NetworkImage(user.photoUrl!),
                       ),
                       title: Text(
@@ -97,15 +106,42 @@ class _CreatePostState extends State<CreatePost> {
                 child: Container(
                   width: MediaQuery.of(context).size.width,
                   height: MediaQuery.of(context).size.width - 30,
-                  decoration: BoxDecoration(
-                    color: Colors.grey[300],
-                    borderRadius: BorderRadius.all(
-                      Radius.circular(5.0),
-                    ),
-                    border: Border.all(
-                      color: Theme.of(context).colorScheme.secondary,
-                    ),
-                  ),
+                  decoration: Theme.of(context).brightness == Brightness.dark
+                      ? BoxDecoration(
+                          boxShadow: [
+                            BoxShadow(
+                              color: Colors.white.withOpacity(0.1),
+                              offset: Offset(-6.0, -6.0),
+                              blurRadius: 16.0,
+                            ),
+                            BoxShadow(
+                              color: Colors.black.withOpacity(0.4),
+                              offset: Offset(6.0, 6.0),
+                              blurRadius: 16.0,
+                            ),
+                          ],
+                          color: Color(0xff2B2B2B),
+                          borderRadius: BorderRadius.circular(12.0),
+                        )
+                      : BoxDecoration(
+                          border: Border.all(
+                            color: Colors.grey.withOpacity(0.1),
+                          ),
+                          boxShadow: [
+                            BoxShadow(
+                              color: Colors.white.withOpacity(0.8),
+                              offset: Offset(-6.0, -6.0),
+                              blurRadius: 16.0,
+                            ),
+                            BoxShadow(
+                              color: Colors.black.withOpacity(0.1),
+                              offset: Offset(6.0, 6.0),
+                              blurRadius: 16.0,
+                            ),
+                          ],
+                          color: Color(0xFFEFEEEE),
+                          borderRadius: BorderRadius.circular(12.0),
+                        ),
                   child: viewModel.imgLink != null
                       ? CustomImage(
                           imageUrl: viewModel.imgLink,
@@ -116,8 +152,9 @@ class _CreatePostState extends State<CreatePost> {
                       : viewModel.mediaUrl == null
                           ? Center(
                               child: Text(
-                                'Upload a Photo',
+                                'Add an Image',
                                 style: TextStyle(
+                                  fontWeight: FontWeight.bold,
                                   color:
                                       Theme.of(context).colorScheme.secondary,
                                 ),
@@ -133,7 +170,7 @@ class _CreatePostState extends State<CreatePost> {
               ),
               SizedBox(height: 20.0),
               Text(
-                'Post Caption'.toUpperCase(),
+                'Post Caption',
                 style: TextStyle(
                   fontSize: 15.0,
                   fontWeight: FontWeight.w600,
@@ -142,7 +179,7 @@ class _CreatePostState extends State<CreatePost> {
               TextFormField(
                 initialValue: viewModel.description,
                 decoration: InputDecoration(
-                  hintText: 'Eg. This is very beautiful place!',
+                  hintText: 'Write a caption',
                   focusedBorder: UnderlineInputBorder(),
                 ),
                 maxLines: null,
@@ -150,7 +187,7 @@ class _CreatePostState extends State<CreatePost> {
               ),
               SizedBox(height: 20.0),
               Text(
-                'Location'.toUpperCase(),
+                'Location',
                 style: TextStyle(
                   fontSize: 15.0,
                   fontWeight: FontWeight.w600,
@@ -160,21 +197,62 @@ class _CreatePostState extends State<CreatePost> {
                 contentPadding: EdgeInsets.all(0.0),
                 title: Container(
                   width: 250.0,
-                  child: TextFormField(
-                    controller: viewModel.locationTEC,
-                    decoration: InputDecoration(
-                      contentPadding: EdgeInsets.all(0.0),
-                      hintText: 'United States,Los Angeles!',
-                      focusedBorder: UnderlineInputBorder(),
+                  height: 40.0,
+                  decoration: Theme.of(context).brightness == Brightness.dark
+                      ? BoxDecoration(
+                          boxShadow: [
+                            BoxShadow(
+                              color: Colors.white.withOpacity(0.1),
+                              blurRadius: 2.0,
+                            ),
+                            BoxShadow(
+                              color: Colors.black.withOpacity(0.4),
+                              offset: Offset(6.0, 6.0),
+                              blurRadius: 10.0,
+                            ),
+                          ],
+                          color: Color(0xff2B2B2B),
+                          borderRadius: BorderRadius.circular(12.0),
+                        )
+                      : BoxDecoration(
+                          border: Border.all(
+                            color: Colors.grey.withOpacity(0.2),
+                          ),
+                          boxShadow: [
+                            BoxShadow(
+                              color: Colors.white.withOpacity(0.8),
+                              offset: Offset(-6.0, -6.0),
+                              blurRadius: 16.0,
+                            ),
+                            BoxShadow(
+                              color: Colors.black.withOpacity(0.1),
+                              offset: Offset(6.0, 6.0),
+                              blurRadius: 10.0,
+                            ),
+                          ],
+                          color: Color(0xFFEFEEEE),
+                          borderRadius: BorderRadius.circular(12.0),
+                        ),
+                  child: Center(
+                    child: Padding(
+                      padding: const EdgeInsets.all(5.0),
+                      child: TextFormField(
+                        controller: viewModel.locationTEC,
+                        decoration: InputDecoration.collapsed(
+                          // contentPadding: EdgeInsets.all(0.0),
+                          hintText: 'United States, Los Angeles!',
+                          // focusedBorder: UnderlineInputBorder(),
+                        ),
+                        maxLines: null,
+                        onChanged: (val) => viewModel.setLocation(val),
+                      ),
                     ),
-                    maxLines: null,
-                    onChanged: (val) => viewModel.setLocation(val),
                   ),
                 ),
                 trailing: IconButton(
                   tooltip: "Use your current location",
                   icon: Icon(
-                    CupertinoIcons.map_pin_ellipse,
+                    Iconsax.location,
                     size: 25.0,
                   ),
                   iconSize: 30.0,
