@@ -1,6 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:social_media_app/auth/register/register.dart';
+import 'package:social_media_app/utils/firebase.dart';
 import 'package:social_media_app/view_models/theme/theme_view_model.dart';
 
 class Setting extends StatefulWidget {
@@ -19,29 +21,28 @@ class _SettingState extends State<Setting> {
           },
           child: Icon(Icons.keyboard_backspace),
         ),
-        backgroundColor: Theme.of(context).backgroundColor,
         elevation: 0.0,
         title: Text(
           "Settings",
           style: TextStyle(),
         ),
       ),
-      backgroundColor: Theme.of(context).backgroundColor,
       body: Padding(
         padding: EdgeInsets.all(10.0),
         child: ListView(
           children: <Widget>[
             ListTile(
-                title: Text(
-                  "About",
-                  style: TextStyle(
-                    fontWeight: FontWeight.w900,
-                  ),
+              title: Text(
+                "About",
+                style: TextStyle(
+                  fontWeight: FontWeight.w900,
                 ),
-                subtitle: Text(
-                  "A Fully Functional Social Media Application Made by CharlyKeleb",
-                ),
-                trailing: Icon(Icons.error)),
+              ),
+              subtitle: Text(
+                "A Fully Functional Social Media Application Made by CharlyKeleb",
+              ),
+              trailing: Icon(Icons.error),
+            ),
             Divider(),
             ListTile(
               title: Text(
@@ -61,6 +62,30 @@ class _SettingState extends State<Setting> {
                 ),
               ),
             ),
+            Divider(),
+            ListTile(
+              onTap: () async {
+                await firebaseAuth.signOut();
+                Navigator.of(context).pushReplacement(
+                  CupertinoPageRoute(
+                    builder: (_) => Register(),
+                  ),
+                );
+              },
+              subtitle: Text(
+                "Logout",
+                style: TextStyle(
+                  fontSize: 22.0,
+                  color: Theme.of(context).iconTheme.color,
+                ),
+              ),
+              trailing: Icon(
+                Icons.chevron_right,
+                size: 30.0,
+                color: Theme.of(context).iconTheme.color,
+              ),
+            ),
+            Divider(),
           ],
         ),
       ),
