@@ -7,7 +7,7 @@ import 'package:social_media_app/utils/firebase.dart';
 abstract class Service {
 
   //function to upload images to firebase storage and retrieve the url.
-  Future<String> uploadImage(Reference ref, File file) async {
+  Future<String> uploadMedia(Reference ref, File file) async {
     String ext = FileUtils.getFileExtension(file);
     Reference storageReference = ref.child("${uuid.v4()}.$ext");
     UploadTask uploadTask = storageReference.putFile(file);
