@@ -1,21 +1,15 @@
 import 'dart:io';
-import 'dart:typed_data';
-
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:image_cropper/image_cropper.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:social_media_app/models/message.dart';
 import 'package:social_media_app/models/status.dart';
-import 'package:social_media_app/models/story_model.dart';
-import 'package:social_media_app/models/user.dart';
+import 'package:social_media_app/posts/create_reel.dart';
 import 'package:social_media_app/posts/story/confrim_status.dart';
 import 'package:social_media_app/services/post_service.dart';
 import 'package:social_media_app/services/status_services.dart';
 import 'package:social_media_app/services/user_service.dart';
 import 'package:social_media_app/utils/constants.dart';
-import 'package:social_media_app/utils/firebase.dart';
 
 class StatusViewModel extends ChangeNotifier {
   //Services
@@ -50,12 +44,11 @@ class StatusViewModel extends ChangeNotifier {
   }
 
   //Functions
-  //Functions
   pickImage({bool camera = false, BuildContext? context}) async {
     loading = true;
     notifyListeners();
     try {
-      PickedFile? pickedFile = await picker.getImage(
+      XFile? pickedFile = await picker.pickImage(
         source: camera ? ImageSource.camera : ImageSource.gallery,
       );
       CroppedFile? croppedFile = await ImageCropper().cropImage(
@@ -85,6 +78,30 @@ class StatusViewModel extends ChangeNotifier {
       Navigator.of(context!).push(
         CupertinoPageRoute(
           builder: (_) => ConfirmStatus(),
+        ),
+      );
+      notifyListeners();
+    } catch (e) {
+      loading = false;
+      notifyListeners();
+      showInSnackBar('Cancelled', context);
+    }
+  }
+
+  //pick video
+  pickVideo({bool camera = false, BuildContext? context}) async {
+    loading = true;
+    notifyListeners();
+    try {
+      XFile? pickedFile = await picker.pickVideo(
+        source: camera ? ImageSource.camera : ImageSource.gallery,
+      );
+
+      mediaUrl = File(pickedFile!.path);
+      loading = false;
+      Navigator.of(context!).push(
+        CupertinoPageRoute(
+          builder: (_) => CreateReel(),
         ),
       );
       notifyListeners();
