@@ -48,130 +48,121 @@ class UserPost extends StatelessWidget {
         onClosed: (v) {},
         closedColor: Theme.of(context).cardColor,
         closedBuilder: (BuildContext context, VoidCallback openContainer) {
-          return Stack(
-            children: [
-              Column(
-                children: [
-                  ClipRRect(
-                    borderRadius: BorderRadius.only(
-                      topLeft: Radius.circular(10.0),
-                      topRight: Radius.circular(10.0),
+          return Padding(
+            padding: const EdgeInsets.all(10.0),
+            child: Stack(
+              children: [
+                Column(
+                  mainAxisAlignment: MainAxisAlignment.start,
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Padding(
+                      padding: const EdgeInsets.only(top: 55.0, bottom: 10),
+                      child: ClipRRect(
+                        borderRadius: BorderRadius.only(
+                          topLeft: Radius.circular(0.0),
+                          topRight: Radius.circular(0.0),
+                        ),
+                        child: CustomImage(
+                          imageUrl: post?.mediaUrl ?? '',
+                          height: 350.0,
+                          fit: BoxFit.cover,
+                          width: double.infinity,
+                        ),
+                      ),
                     ),
-                    child: CustomImage(
-                      imageUrl: post?.mediaUrl ?? '',
-                      height: 350.0,
-                      fit: BoxFit.cover,
-                      width: double.infinity,
+                    Visibility(
+                      visible: post!.description != null &&
+                          post!.description.toString().isNotEmpty,
+                      child: Padding(
+                        padding: const EdgeInsets.only(left: 0.0, top: 3.0),
+                        child: Text(
+                          '${post?.description ?? "".trim()}',
+                          style: TextStyle(
+                            fontSize: 15.0,
+                          ),
+                          maxLines: 2,
+                        ),
+                      ),
                     ),
-                  ),
-                  Padding(
-                    padding:
-                        EdgeInsets.symmetric(horizontal: 3.0, vertical: 5.0),
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.start,
-                      crossAxisAlignment: CrossAxisAlignment.start,
+                    SizedBox(height: 5.0),
+                    Row(
                       children: [
-                        Padding(
-                          padding: const EdgeInsets.only(left: 0.0),
-                          child: Row(
-                            children: [
-                              buildLikeButton(),
-                              SizedBox(width: 5.0),
-                              InkWell(
-                                borderRadius: BorderRadius.circular(10.0),
-                                onTap: () {
-                                  Navigator.of(context).push(
-                                    CupertinoPageRoute(
-                                      builder: (_) => Comments(post: post),
-                                    ),
-                                  );
-                                },
-                                child: Icon(
-                                  CupertinoIcons.chat_bubble,
-                                  size: 25.0,
-                                ),
+                        InkWell(
+                          borderRadius: BorderRadius.circular(10.0),
+                          onTap: () {
+                            Navigator.of(context).push(
+                              CupertinoPageRoute(
+                                builder: (_) => Comments(post: post),
                               ),
-                            ],
+                            );
+                          },
+                          child: Icon(
+                            Ionicons.chatbox_ellipses,
+                            size: 25.0,
                           ),
                         ),
-                        SizedBox(height: 5.0),
-                        Row(
-                          mainAxisAlignment: MainAxisAlignment.start,
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            Flexible(
-                              child: Padding(
-                                padding: const EdgeInsets.only(left: 0.0),
-                                child: StreamBuilder(
-                                  stream: likesRef
-                                      .where('postId', isEqualTo: post!.postId)
-                                      .snapshots(),
-                                  builder: (context,
-                                      AsyncSnapshot<QuerySnapshot> snapshot) {
-                                    if (snapshot.hasData) {
-                                      QuerySnapshot snap = snapshot.data!;
-                                      List<DocumentSnapshot> docs = snap.docs;
-                                      return buildLikesCount(
-                                          context, docs.length ?? 0);
-                                    } else {
-                                      return buildLikesCount(context, 0);
-                                    }
-                                  },
-                                ),
-                              ),
-                            ),
-                            SizedBox(width: 5.0),
-                            StreamBuilder(
-                              stream: commentRef
-                                  .doc(post!.postId!)
-                                  .collection("comments")
-                                  .snapshots(),
-                              builder: (context,
-                                  AsyncSnapshot<QuerySnapshot> snapshot) {
-                                if (snapshot.hasData) {
-                                  QuerySnapshot snap = snapshot.data!;
-                                  List<DocumentSnapshot> docs = snap.docs;
-                                  return buildCommentsCount(
-                                      context, docs.length ?? 0);
-                                } else {
-                                  return buildCommentsCount(context, 0);
-                                }
-                              },
-                            ),
-                          ],
+                        SizedBox(width: 5.0),
+                        StreamBuilder(
+                          stream: commentRef
+                              .doc(post!.postId!)
+                              .collection("comments")
+                              .snapshots(),
+                          builder:
+                              (context, AsyncSnapshot<QuerySnapshot> snapshot) {
+                            if (snapshot.hasData) {
+                              QuerySnapshot snap = snapshot.data!;
+                              List<DocumentSnapshot> docs = snap.docs;
+                              return buildCommentsCount(
+                                  context, docs.length ?? 0);
+                            } else {
+                              return buildCommentsCount(context, 0);
+                            }
+                          },
                         ),
-                        Visibility(
-                          visible: post!.description != null &&
-                              post!.description.toString().isNotEmpty,
-                          child: Padding(
-                            padding: const EdgeInsets.only(left: 5.0, top: 3.0),
-                            child: Text(
-                              '${post?.description ?? ""}',
-                              style: TextStyle(
-                                color:
-                                    Theme.of(context).textTheme.caption!.color,
-                                fontSize: 15.0,
-                              ),
-                              maxLines: 2,
-                            ),
-                          ),
+                        Spacer(),
+                        buildLikeButton(),
+                        SizedBox(width: 5.0),
+                        StreamBuilder(
+                          stream: likesRef
+                              .where('postId', isEqualTo: post!.postId)
+                              .snapshots(),
+                          builder:
+                              (context, AsyncSnapshot<QuerySnapshot> snapshot) {
+                            if (snapshot.hasData) {
+                              QuerySnapshot snap = snapshot.data!;
+                              List<DocumentSnapshot> docs = snap.docs;
+                              return buildLikesCount(context, docs.length ?? 0);
+                            } else {
+                              return buildLikesCount(context, 0);
+                            }
+                          },
                         ),
-                        SizedBox(height: 3.0),
-                        Padding(
-                          padding: const EdgeInsets.all(3.0),
-                          child: Text(
-                            timeago.format(post!.timestamp!.toDate()),
-                            style: TextStyle(fontSize: 10.0),
-                          ),
-                        ),
-                        // SizedBox(height: 5.0),
                       ],
                     ),
-                  )
-                ],
-              ),
-              buildUser(context),
-            ],
+                    Padding(
+                      padding:
+                          EdgeInsets.symmetric(horizontal: 3.0, vertical: 5.0),
+                      child: Column(
+                        mainAxisAlignment: MainAxisAlignment.start,
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        children: [
+                          Padding(
+                            padding: const EdgeInsets.all(3.0),
+                            child: Text(
+                              timeago.format(post!.timestamp!.toDate()),
+                              style: TextStyle(fontSize: 10.0),
+                            ),
+                          ),
+                          // SizedBox(height: 5.0),
+                        ],
+                      ),
+                    )
+                  ],
+                ),
+                buildUser(context),
+              ],
+            ),
           );
         },
       ),
@@ -188,31 +179,6 @@ class UserPost extends StatelessWidget {
         if (snapshot.hasData) {
           List<QueryDocumentSnapshot> docs = snapshot.data?.docs ?? [];
 
-          ///replaced this with an animated like button
-          // return IconButton(
-          //   onPressed: () {
-          //     if (docs.isEmpty) {
-          //       likesRef.add({
-          //         'userId': currentUserId(),
-          //         'postId': post!.postId,
-          //         'dateCreated': Timestamp.now(),
-          //       });
-          //       addLikesToNotification();
-          //     } else {
-          //       likesRef.doc(docs[0].id).delete();
-          //       services.removeLikeFromNotification(
-          //           post!.ownerId!, post!.postId!, currentUserId());
-          //     }
-          //   },
-          //   icon: docs.isEmpty
-          //       ? Icon(
-          //           CupertinoIcons.heart,
-          //         )
-          //       : Icon(
-          //           CupertinoIcons.heart_fill,
-          //           color: Colors.red,
-          //         ),
-          // );
           Future<bool> onLikeButtonTapped(bool isLiked) async {
             if (docs.isEmpty) {
               likesRef.add({
@@ -236,10 +202,11 @@ class UserPost extends StatelessWidget {
             circleColor:
                 CircleColor(start: Color(0xffFFC0CB), end: Color(0xffff0000)),
             bubblesColor: BubblesColor(
-                dotPrimaryColor: Color(0xffFFA500),
-                dotSecondaryColor: Color(0xffd8392b),
-                dotThirdColor: Color(0xffFF69B4),
-                dotLastColor: Color(0xffff8c00)),
+              dotPrimaryColor: Color(0xffFFA500),
+              dotSecondaryColor: Color(0xffd8392b),
+              dotThirdColor: Color(0xffFF69B4),
+              dotLastColor: Color(0xffff8c00),
+            ),
             likeBuilder: (bool isLiked) {
               return Icon(
                 docs.isEmpty ? Ionicons.heart_outline : Ionicons.heart,
@@ -278,12 +245,12 @@ class UserPost extends StatelessWidget {
 
   buildLikesCount(BuildContext context, int count) {
     return Padding(
-      padding: const EdgeInsets.only(left: 7.0),
+      padding: const EdgeInsets.only(left: 0.0, right: 5.0),
       child: Text(
         '$count likes',
         style: TextStyle(
           fontWeight: FontWeight.bold,
-          fontSize: 10.0,
+          fontSize: 12.5,
         ),
       ),
     );
@@ -293,8 +260,8 @@ class UserPost extends StatelessWidget {
     return Padding(
       padding: const EdgeInsets.only(top: 0.5),
       child: Text(
-        '-   $count comments',
-        style: TextStyle(fontSize: 8.5, fontWeight: FontWeight.bold),
+        ' $count  comments',
+        style: TextStyle(fontSize: 12.5, fontWeight: FontWeight.bold),
       ),
     );
   }
@@ -308,73 +275,68 @@ class UserPost extends StatelessWidget {
           DocumentSnapshot snap = snapshot.data!;
           UserModel user =
               UserModel.fromJson(snap.data() as Map<String, dynamic>);
-          return Visibility(
-            visible: !isMe,
-            child: Align(
-              alignment: Alignment.topCenter,
-              child: Container(
-                height: 50.0,
-                decoration: BoxDecoration(
-                  color: Colors.white60,
-                  borderRadius: BorderRadius.only(
-                    topLeft: Radius.circular(10.0),
-                    topRight: Radius.circular(10.0),
-                  ),
-                ),
-                child: GestureDetector(
-                  onTap: () => showProfile(context, profileId: user.id!),
-                  child: Padding(
-                    padding: const EdgeInsets.only(left: 10.0),
-                    child: Row(
-                      crossAxisAlignment: CrossAxisAlignment.center,
+          return Align(
+            alignment: Alignment.topCenter,
+            child: Container(
+              height: 50.0,
+              // decoration: BoxDecoration(
+              //   color: Colors.white60,
+              //   borderRadius: BorderRadius.only(
+              //     topLeft: Radius.circular(10.0),
+              //     topRight: Radius.circular(10.0),
+              //   ),
+              // ),
+              child: GestureDetector(
+                onTap: () => showProfile(context, profileId: user.id!),
+                child: Row(
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: [
+                    user.photoUrl!.isEmpty
+                        ? CircleAvatar(
+                            radius: 20.0,
+                            backgroundColor:
+                                Theme.of(context).colorScheme.secondary,
+                            child: Center(
+                              child: Text(
+                                '${user.username![0].toUpperCase()}',
+                                style: TextStyle(
+                                  color: Colors.white,
+                                  fontSize: 15.0,
+                                  fontWeight: FontWeight.w900,
+                                ),
+                              ),
+                            ),
+                          )
+                        : CircleAvatar(
+                            radius: 20.0,
+                            backgroundImage: CachedNetworkImageProvider(
+                              '${user.photoUrl}',
+                            ),
+                          ),
+                    SizedBox(width: 5.0),
+                    Column(
+                      mainAxisSize: MainAxisSize.min,
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      mainAxisAlignment: MainAxisAlignment.center,
                       children: [
-                        user.photoUrl!.isEmpty
-                            ? CircleAvatar(
-                                radius: 20.0,
-                                backgroundColor:
-                                    Theme.of(context).colorScheme.secondary,
-                                child: Center(
-                                  child: Text(
-                                    '${user.username![0].toUpperCase()}',
-                                    style: TextStyle(
-                                      color: Colors.white,
-                                      fontSize: 15.0,
-                                      fontWeight: FontWeight.w900,
-                                    ),
-                                  ),
-                                ),
-                              )
-                            : CircleAvatar(
-                                radius: 20.0,
-                                backgroundImage: CachedNetworkImageProvider(
-                                  '${user.photoUrl}',
-                                ),
-                              ),
-                        SizedBox(width: 5.0),
-                        Column(
-                          mainAxisSize: MainAxisSize.min,
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            Text(
-                              '${post?.username ?? ""}',
-                              style: TextStyle(
-                                fontWeight: FontWeight.w900,
-                                color: Colors.black,
-                              ),
-                              overflow: TextOverflow.ellipsis,
-                            ),
-                            Text(
-                              '${post?.location ?? 'Wooble'}',
-                              style: TextStyle(
-                                fontSize: 10.0,
-                                color: Color(0xff4D4D4D),
-                              ),
-                            ),
-                          ],
+                        Text(
+                          '${post?.username?.trim() ?? ""}',
+                          style: TextStyle(
+                            fontWeight: FontWeight.w900,
+                            // color: Colors.black,
+                          ),
+                          overflow: TextOverflow.ellipsis,
+                        ),
+                        Text(
+                          '${post?.location?.trim() ?? 'Wooble'}',
+                          style: TextStyle(
+                            fontSize: 10.0,
+                            // color: Color(0xff4D4D4D),
+                          ),
                         ),
                       ],
                     ),
-                  ),
+                  ],
                 ),
               ),
             ),
